chore(tests): split jetstream tests
This reorganization reduces the amount of tests in each file for
clarity, and simplifies launching them separately. This is necessary
due to a memory leak that happens in Jetstream Pytorch when switching
models.
